refactor(routing): extract routes configuration into separate file
